MT#60197 support deleting call recordings per subscriber
admin panel will no longer delete callrecordings for
all parties of the call, but the single subscriber only.

rest-api, behaves the same way, if the ?subscriber_id=
parameter is provided.

when a callrecording is deleted by all call parties,
the remaining records and recording files will be
dropped.

(cherry picked from commit 1c09e2773a4ff1c441f9dc4b86e1dcd15cad4f39)
(cherry picked from commit 22c5b8ad7efc7428c932327b04961c820b9f371b)

+++ lib/NGCP/Panel/Utils/Subscriber.pm
@@ -2522,23 +2522,43 @@ sub get_voicemail_content_type {
 
 sub delete_callrecording {
     my %params = @_;
-    my($recording, $force_delete) = @params{qw/recording force_delete/};
+    my($c,$recording, $force_delete, $uuid) = @params{qw/c recording force_delete uuid/};
 
-    foreach my $stream($recording->recording_streams->all) {
-        #if we met some error deleting file - we will fail and transaction will be rollbacked
-        if (! -e $stream->full_filename && !$force_delete) {
-            die "Callrecording file ".$stream->full_filename." is absent";
-        }
-        eval {
-            unlink $stream->full_filename;
-        };
-        if ($@ && !$force_delete) {
-            die("Cannot delete call recording file: $@");
+    $recording = $c->model('DB')->resultset('recording_calls')->find({
+        id => $recording->id
+    },{for => 'update'});
+
+    my $delete_all = 1;
+    if ($uuid) {
+        $recording->recording_metakeys->search_rs({
+            'key' => 'uuid',
+            'value' => $uuid,
+        })->delete;
+        if ($recording->recording_metakeys->search_rs({
+                'key' => 'uuid',
+            })->first) {
+            $delete_all = 0;
         }
     }
-    $recording->recording_streams->delete;
-    $recording->recording_metakeys->delete;
-    $recording->delete;
+
+    if ($delete_all) {
+        foreach my $stream($recording->recording_streams->all) {
+            #if we met some error deleting file - we will fail and transaction will be rollbacked
+            if (! -e $stream->full_filename && !$force_delete) {
+                die "Callrecording file ".$stream->full_filename." is absent";
+            }
+            eval {
+                unlink $stream->full_filename;
+            };
+            if ($@ && !$force_delete) {
+                die("Cannot delete call recording file: $@");
+            }
+        }
+        $recording->recording_metakeys->delete;
+        $recording->recording_streams->delete;
+        $recording->delete;
+    }
+    
 }
 
 sub prov_to_billing_subscriber_id {
